backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0
python-dotenv>=1.0.0
slowapi>=0.1.9
supabase>=2.0.0
anthropic>=0.40.0
httpx>=0.27.0
